# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Task management service with teams and change history.",
        "title": "Task Tracker API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/analytics/teams": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Team stats: members and tasks done in the last 7 days (admin only)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_transport_http_analytics.teamStatsResponse"
                            }
                        }
                    },
                    "403": {
                        "description": "not an admin",
                        "schema": {
                            "$ref": "#/definitions/task_tracker_internal_transport_http_httpkit.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authenticate and issue a JWT",
                "parameters": [
                    {
                        "description": "email and password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_auth.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_auth.loginResponse"
                        }
                    },
                    "400": {
                        "description": "invalid json",
                        "schema": {
                            "$ref": "#/definitions/task_tracker_internal_transport_http_httpkit.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/task_tracker_internal_transport_http_httpkit.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_auth.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_auth.userResponse"
                        }
                    },
                    "400": {
                        "description": "invalid data",
                        "schema": {
                            "$ref": "#/definitions/task_tracker_internal_transport_http_httpkit.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "email already taken",
                        "schema": {
                            "$ref": "#/definitions/task_tracker_internal_transport_http_httpkit.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/teams": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Teams the current user belongs to",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_transport_http_teams.teamResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "missing token",
                        "schema": {
                            "$ref": "#/definitions/task_tracker_internal_transport_http_httpkit.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Create a team, creator becomes owner",
                "parameters": [
                    {
                        "description": "team name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_teams.createTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_teams.teamResponse"
                        }
                    },
                    "400": {
                        "description": "invalid data",
                        "schema": {
                            "$ref": "#/definitions/task_tracker_internal_transport_http_httpkit.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "missing token",
                        "schema": {
                            "$ref": "#/definitions/task_tracker_internal_transport_http_httpkit.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/teams/{id}/invite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Invite a user to a team (team owner/admin or global admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "team id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "invitee email",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_teams.inviteRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "invalid data",
                        "schema": {
                            "$ref": "#/definitions/task_tracker_internal_transport_http_httpkit.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "$ref": "#/definitions/task_tracker_internal_transport_http_httpkit.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "team or user not found",
                        "schema": {
                            "$ref": "#/definitions/task_tracker_internal_transport_http_httpkit.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "already a member",
                        "schema": {
                            "$ref": "#/definitions/task_tracker_internal_transport_http_httpkit.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_transport_http_analytics.teamStatsResponse": {
            "type": "object",
            "properties": {
                "done_last_7d": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "members": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "internal_transport_http_auth.loginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http_auth.loginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http_auth.registerRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http_auth.userResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "internal_transport_http_teams.createTeamRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http_teams.inviteRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http_teams.teamResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "task_tracker_internal_transport_http_httpkit.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT as: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}